requirements.txt
streamlit
langchain
langchain-openai
psycopg2-binary
python-dotenv